libcontainer: specconv: factor out systemd specific functions
Factor out an init function specific to systemd, so subsequent patches
can easily make building systemd support build-time optional.

[@kolyshkin: wording]
